Move Struct and Union classes to separate files. StructOrUnion -> Record
