+++ bindgen/TypeTranslator.h
@@ -35,10 +35,9 @@ class TypeTranslator {
      */
     std::map<std::string, std::string> typeMap;
 
-    std::shared_ptr<Type>
-    translateStructOrUnionOrEnum(const clang::QualType &qtpe);
+    std::shared_ptr<Type> translateRecordOrEnum(const clang::QualType &qtpe);
 
-    std::shared_ptr<Type> translateStructOrUnion(const clang::QualType &qtpe);
+    std::shared_ptr<Type> translateRecord(const clang::QualType &qtpe);
 
     std::shared_ptr<Type> translateFunctionPointer(const clang::QualType &qtpe);
